Change location of doc mapper and add it to the migration
