+++ build.gradle
@@ -276,7 +276,7 @@ dependencyManagement {
 dependencies {
 
     implementation group: 'com.github.hmcts', name: 'ecm-common', version: '2.1.2'
-    implementation group: 'com.github.hmcts', name: 'ecm-data-model', version: '1.13.14'
+    implementation group: 'com.github.hmcts', name: 'ecm-data-model', version: '1.3.14'
     implementation group: 'com.github.hmcts', name: 'et-data-model', version: '3.18.30'
     implementation group: 'com.github.hmcts', name: 'ccd-case-document-am-client', version: '1.7.3'
     implementation group: 'com.github.hmcts', name: 'document-management-client', version: '7.0.1'
@@ -326,7 +326,7 @@ dependencies {
     annotationProcessor 'org.projectlombok:lombok'
 
     testImplementation group: 'com.auth0', name: 'java-jwt', version: '4.5.0'
-    testImplementation 'com.github.hmcts:fortify-client:1.4.7:all'
+    testImplementation 'com.github.hmcts:fortify-client:1.4.8:all'
     testImplementation group: 'com.warrenstrange', name: 'googleauth', version: '1.5.0'
     testImplementation group: 'javax.xml.bind', name: 'jaxb-api', version: '2.3.1'
     testImplementation group: 'net.serenity-bdd', name: 'serenity-core', version: serenity
