+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/AddDocumentControllerTest.java
@@ -10,29 +10,22 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import uk.gov.hmcts.ecm.common.model.ccd.SubmitEvent;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseDetails;
-import uk.gov.hmcts.ethos.replacement.docmosis.DocmosisApplication;
-import uk.gov.hmcts.ethos.replacement.docmosis.service.DocumentManagementService;
+import uk.gov.hmcts.ecm.common.model.ccd.SubmitEvent;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.VerifyTokenService;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.JsonMapper;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.util.Objects;
 
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -46,8 +39,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
     private static final String AUTH_TOKEN = "Bearer eyJhbGJbpjciOiJIUzI1NiJ9";
     private static final String ABOUT_TO_SUBMIT_URL = "/addDocument/aboutToSubmit";
 
-    @MockBean
-    private DocumentManagementService documentManagementService;
     @MockBean
     private VerifyTokenService verifyTokenService;
     private MockMvc mvc;
@@ -77,7 +68,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
                 .andExpect(jsonPath(JsonMapper.DATA, notNullValue()))
                 .andExpect(jsonPath(JsonMapper.ERRORS, nullValue()))
                 .andExpect(jsonPath(JsonMapper.WARNINGS, nullValue()));
-        verify(documentManagementService, times(1)).addUploadedDocsToCaseDocCollection(any());
     }
 
     @Test
@@ -88,27 +78,11 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
                         .header(HttpHeaders.AUTHORIZATION, AUTH_TOKEN)
                         .content(requestContent.toString()))
                 .andExpect(status().isForbidden());
-        verify(documentManagementService, never()).addUploadedDocsToCaseDocCollection(any());
-    }
-
-    @Test
-    void testAboutToSubmitWithInvalidDocumentIndex() throws Exception {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-        doThrow(new RuntimeException("Invalid document index"))
-                .when(documentManagementService).addUploadedDocsToCaseDocCollection(any());
-
-        mvc.perform(post(ABOUT_TO_SUBMIT_URL)
-                        .content(requestContent.toString())
-                        .header("Authorization", AUTH_TOKEN)
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath(JsonMapper.ERRORS, notNullValue()));
-        verify(documentManagementService, times(1)).addUploadedDocsToCaseDocCollection(any());
     }
 
     private void doRequestSetUp() throws IOException, URISyntaxException {
         ObjectMapper objectMapper = new ObjectMapper();
-        requestContent = objectMapper.readTree(new File(getClass()
-                .getResource("/exampleV1.json").toURI()));
+        requestContent = objectMapper.readTree(new File(Objects.requireNonNull(getClass()
+                .getResource("/exampleV1.json")).toURI()));
     }
 }
